requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
anthropic>=0.39.0
jinja2>=3.1.0
numpy>=1.24.0
google-search-results>=2.4.2
playwright>=1.40.0
